minio client: retry on some errors (#1880)
## Context

In #1874 we made a hotfix that would fix a bug where Grist would
override an existing document with an empty one when the
MinIOExternalStorage client was encountering ECONNRESET or
InternalErrors. @paulfitz suggested that it would be even better if the
client would retry on those errors.

## Proposed solution

Teach MinIOExternalStorage about what a true FatalError is, and let the
ChecksummedExternalStorage handle the retries.

## Has this been tested?

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help -- I haven't seen much retrying logic
being tested in the codebase so far

---------

Co-authored-by: Florent <florent.fayolle+git@beta.gouv.fr>