(core) add a link to help for suggestions feature
Summary:
This adds a link to https://support.getgrist.com/sharing/#suggestions
for the suggestions feature. It is present on a tooltip for the
button that turns the feature on, and on the suggestions page
itself in its initial empty state.

The linked section was added as a placeholder for suggestions documentation
in https://github.com/gristlabs/grist-help/pull/568/files

Test Plan: manual

Reviewers: Spoffy

Reviewed By: Spoffy

Differential Revision: https://phab.getgrist.com/D4758