pruneAPIResult: don't prune `disabledAt`
On second thought (re: c49a6124866ba00bb5c70c3ac6c90dc7d8044045) I do
want to have `disabledAt` showing through the API. Originally, there
was no need to show `disabledAt` for users, but there is no harm in
showing when a user was disabled. I now need this property on
documents, so I can use it in the UI to indicate if a document is
disabled or not.