fix: Add missing login in Service Accounts API Responses (#1919)
## Context

Testing the 1.7.5 we discover that we can't complete the following user
story:

"As a user I want to create a service account and add it to a document"
The Service account API do not expose logins, so it's hard to add them
to any doc, workspace or org.

## Proposed solution

The login is mandatory to add a service account to a document, a
workspace
or an organization.
This PR fixes that miss by adding a login key to POST, GET, PATCH,
DELETE routes.

It also checks api data responses for each routes.
It modifies the tests accordingly.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [x] 👍 yes, I added tests to the test suite

The test using this behavior is now using API response to get the
service account login rather than querying the db.

---------

Co-authored-by: fflorent <florent.fayolle+git@beta.gouv.fr>