show a count of changes when making a suggestion
This adds a visible count of changes when the user is
preparing a suggestion. The count is not very meaningful,
being just the number of actions in the action stack,
but the main point is to draw the eye towards the area
where suggestions are summarized and can be finalized.