make suggestions work better if target has had table/column renames (#1918)
Add tests for various possible changes of the target document, and add
support for accepting suggestions successfully in those cases.

We "rebase" the saved suggestion against any column/table renames made
since it was created.

Includes some work to make the arguments of `concatenateSummary`
read-only. For simplicity, it used to modify its arguments, but that
means deep copies have to be done by callers.