Allows GRIST_SANDBOX to override Pyodide node binary (#1932)
## Context

When Grist is packed into Flatpak, everything runs inside the Flatpak
sandbox, and our main electron app is invoked through the "Zypak"
wrapper. This wrapper does weird syscall interceptions and overrides via
LD_PRELOAD to magically make Chromium's sandbox work inside the Flatpak
sandbox.

The core issue seems to have been we're trying to re-invoke the
grist-desktop binary (electron) inside the Zypak wrapper. Something in
the wrapper is causing it to throw ENOENT.

## This commit

This allows Grist Desktop to workaround this issue by changing the
NSandbox approach for Pyodide. It:
- Makes it use GRIST_SANDBOX env var value as a binary
- Passes ELECTRON_RUN_AS_NODE as an environment variable to force
Electron to behave as Node.js
- Allows arguments to be passed to GRIST_SANDBOX