Forms: make sure the Submit row can't be removed even w/ keyboard
In the form edit view, while we didn't show the trash can icon to remove
the submit button when hovering it (good), the submit model wasn't
actually prevented from being removed in other ways (bad).

When we use keyboard to tab through all form fields, at one point we are
focused on the submit button. The focus ring isn't shown, but we are
focused on it. Pressing Backspace deletes the submit button and there is
no way to retrieve it.

This prevents submit button deletion with the keyboard