choices text box: make sure all choice tokens are readable
This is a follow-up to
https://github.com/gristlabs/grist-core/pull/1529,
where we made sure that choice tokens with custom bg color but no custom
fg color would not become unreadable when switching from light to dark
themes.

The tokens that were rendered when editing a text cell did not benefit
from this fix. So while the tokens stayed readable when looking at them,
they could go back to being unreadable when editing them after
double-clicking the cell (with automatic light text in dark theme, on a
light custom bg, for example).

This fixes that use-case.