Bump @types/underscore to 1.11.15 (#1980)
## Context

1. @types/underscore's recent versions introduce a generic variant of
`_.unzip`, which causes the existing type inference to break.
2. We're using @types/underscore implicitly from transitive dependencies
(despite using underscore directly).

This issue would be hit eventually when the version is updated. This
case has already been in hit in Grist Desktop, which uses
@types/underscore 1.11.15 instead of 1.11.0, preventing core
compilation.

## This commit

This adds @types/underscore at 1.11.15 as an explicit dependency, and
updates `GridView.ts` to compile with the new type signatures for
`_.unzip`.