Show a "record is unavailable" message when a user opens a record card blocked by ACLs (#1896)
## Context

When a record card is opened for a row that is unavailable (blocked by
access control), currently it shows a "random" record.

This occurs because Cursor refuses to be set to non-existent row ids,
and instead maintains the current rowIndex. This is usually the last one
viewed, or rowIndex 0.

## This commit

`Cursor` supports `rowIndex` being set to null currently, but this
generally doesn't occur due to `Cursor` clamping the row indexes.

This uses the Cursor's null rowIndex as an 'Invalid record selected'
state, and updates the record view to show a message when the rowIndex
is null.