(core) Fix a situation that sometimes causes excessive memory use in the frontend.
Summary:
It's easiest to explain with an example. Let's say you have 100 views
for a number of tables, each with 20 column. We create an in-memory
model for each view, which includes an observable array of all the
available columns. Normally, that would be O(100*20) entities. Now, imagine that
the user opening the document is limited by access rules to seeing only
one table/view. Nearly all of the 100*20 columns are now censored. That
sets their `parentId` (which table they belong to) to 0.

What we did previously was associate these columns with a default "empty
record", a RowModel representing the non-existent 0 table record. This RowModel
would then contain an observable array of 2k columns (since they are censored,
they look like they all belong to table 0). The real issue is that there are 100
censored views which would all use this placeholder RowModel, and build extra
state for those 2k columns: that's 100*2k = 200k entities.

With this fix, the observable arrays for the "empty records" will always
be empty, ignoring any items that may have a parentId of 0.

Test Plan:
Tested manually with memory tools in Firefox and Chrome. On one document,
Chrome's snapshot came in at 957MB before the fix and 167MB after.

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4805