(core) Optimize GranularAccess _pruneRows method.
Summary:
It was previously going through every column in every ruleSet. This is
wasteful when a ruleSet has many columns, since the cell-censoring
decisions depend only on the ruleSet itself. With this refactor, we make
the decisions based on the ruleSet once for all the columns it covers.

Also expanded a todo about another place that's likely to cause performance
problems.

Test Plan:
Relying on existing tests to test correctness.
For timing, timed manually: for one sample document with a rule containing 300
columns, time spent processing _pruneRows went from ~30s to 0.7s.

Reviewers: georgegevoian, jarek, Spoffy

Reviewed By: georgegevoian, Spoffy

Subscribers: Spoffy, jarek

Differential Revision: https://phab.getgrist.com/D4807