Eslint: fix @typescript-eslint/no-empty-object-type
-> app/client/lib/GristWindow.ts:
GristWindow is exported but never used. Let's instead overwrite the
Window type.

-> app/client/ui/MultiSelector.ts:
selectOptions had the wrong type, a caller passed an object of this
form: `{ defLabel: string }`

-> app/server/lib/ExportTableSchema.ts:
There has been an issue with the types of the functions, with a typo for
the `constraints` property (in singular in the interface but return in
plural in functions).
Let's extract an FrictionlessFormat interface and use it for the return
type of the `buildTypeField` function

-> test/server/lib/DocWorkerLoadTracker.ts:
confusion between `() => {}` and `() => void` (the first one is a
function expression returning nothing, the second is the type of the
latter)