Eslint: Remove function appCommWaiter (dead code) (#2001)
## Context

I open this PR while working on #2000. Eslint reports an accurate issue
with `waitForActiveRequests()`.

While digging, I found that `appCommWaiter()` function had a
implementation problem: it used `waitForActiveRequests()` which
supposedly awaited for pending request to end. But the object awaited
was not a promise and the result was resolved immediately. Also this
function does not seem to be used elsewhere (at least from what I have
found in grist-core).

## Proposed solution

1. Remove `appCommWaiter` which does not seem to be used.
2. Remove `waitForActiveRequests` which was only used in the above
function

## Related issues

#1967 

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

## Screenshots / Screencasts

<!-- delete if not relevant -->