(core) `Owner` permission is now needed for managing orgs and workspaces
Summary:
Restrict org and workspace modifications to owners only. This includes:
- renaming
- updating preferences for all users in org
- removing
- moving to trash and restoring.

Code actually checks if user is an owner by testing:
- For updates: SCHEMA_EDIT
- For removals: SCHEMA_EDIT + REMOVE
This will be more flexible for future.

This correlates to an existing behavior for documents, which can be
renamed, moved or deleted only by owners.

Test Plan: Updated existing tests.

Reviewers: dsagal, paulfitz

Reviewed By: dsagal, paulfitz

Subscribers: paulfitz, dsagal

Differential Revision: https://phab.getgrist.com/D4800