Fix how zero Date/DateTimes look to other formulas (#1995)
## Context

The date 1970-01-01 is incorrectly cast as a float in formula columns.
See #1994 for more details.

Resolves #1994 

## Proposed solution

Date/DateTime values are stored as floats. To check if it should be
presented as a date/datetime object, we should check if we have a float
rather than if the value is truthy (as that incorrectly interprets zeros
as non-dates).

## Has this been tested?

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->