(core) Store OIDCServer entities in Redis/TypeORM and add clients API
Summary:
This builds on [D4746](https://phab.getgrist.com/D4746) by implementing `oidc-provider`
adapters for both Redis and TypeORM. Entities are stored in one or the other depending
on their nature (e.g. transient entities like interactions and sessions are stored using
the Redis backend, while other entities like OAuth clients and grants are stored using the
TypeORM backend for durability and referential integrity. `REDIS_URL` must be set for the
new adapters to be used in place of the default in-memory adapter, which remains as a fallback.

Clients may now be registered using the REST API. The endpoints are:
 - `GET /oauth/clients`: List all clients
 - `GET /oauth/clients/:id`: Get the client with the specified ID
 - `POST /oauth/clients`: Create a new client
 - `DELETE /oauth/clients/:id`: Delete the client with the specified ID

For example, the following request registers a new client with name "ACME Corp" and redirect
URI http://localhost:8484/oauth2/callback:

```
curl -X POST http://localhost:8080/o/docs/api/oauth/clients \
  -H "Authorization: Bearer GRIST_API_KEY" \
  -H "Content-Type: application/json" \
  -d '{"name":"ACME Corp","redirectUri":"http://localhost:8484/oauth2/callback"}'
```

The POST response includes the secret, which is not exposed by the GET endpoints.

Client entities are owned by the request user, and the endpoints above only apply to clients
associated with that user.

The `GRIST_OIDC_SERVER_CLIENTS` variable has been replaced with the new clients API.
The API will be used in the future for a new UI flow that allows users on getgrist.com to
register new clients for a new login method for self-managed Grist.

Test Plan: Browser and server tests

Reviewers: jarek, paulfitz, dsagal

Reviewed By: jarek, dsagal

Subscribers: dsagal

Differential Revision: https://phab.getgrist.com/D4756