be somewhat robust if table/columns can't be related to current state
History may be truncated, so we may not know how the table/columns
in a suggestion relate to the table/columns in the current doc.
Should be able to at least show the suggestion in some form, even if
it can't be applied.

Separately, an important improvement would be when table/columns
CAN be related after renames, that should be used in rendering.
This will be done in follow-up work.