show a count of changes when making a suggestion (#1960)
This adds a visible count of changes when the user is preparing a
suggestion. The point of this is not so much the count itself, as to
draw the eye towards the area where suggestions are summarized and can
be finalized.

<img width="1119" height="545" alt="image"
src="https://github.com/user-attachments/assets/2b3b1c33-da01-4d27-8150-4b9e4f102864"
/>