Fixing auth provider tests (#2017)
## Context

New auth provider tests are not prepared for the GRIST_TEST_LOGIN
provider, as this Grist reports this as an error, as this provider is
not listed in list of supported provider list.

## Proposed solution

Current tests are using GRIST_TEST_LOGIN provider which is not needed
for them to work.

## Has this been tested?

Existing test should pass. I also tested `saas` flavor and the failing
test was passing.

PR: https://github.com/gristlabs/grist-core/pull/2017