(core) Hide 'Create a form' option on Raw Data sections, which leads to an error.
Summary:
The error is particularly confusing: "[Sandbox] TypeError '<' not
supported between instances of 'int' and 'str'". Ultimately, it's caused
because we try to pass the raw-data page's name ("data") as if it were a
viewRef to CreateViewSection action.

Test Plan: Added a test case to ensure the item is hidden on raw data pages.

Reviewers: paulfitz

Reviewed By: paulfitz

Subscribers: paulfitz

Differential Revision: https://phab.getgrist.com/D4824